+++ test/Tier/CombineTier.sol.ts
@@ -1,5 +1,5 @@
 import * as Util from "../../utils";
-import chai from "chai";
+import { assert } from "chai";
 import { ethers } from "hardhat";
 import { concat, hexlify } from "ethers/lib/utils";
 import { op, paddedUInt32, paddedUInt256 } from "../../utils";
@@ -7,8 +7,6 @@ import type { Contract } from "ethers";
 
 import type { CombineTier } from "../../typechain/CombineTier";
 import type { ReadWriteTier } from "../../typechain/ReadWriteTier";
-
-const { assert } = chai;
 
 enum Tier {
   ZERO,
