+++ test/Tier/ReadWriteTier.sol.ts
@@ -1,4 +1,4 @@
-import chai from "chai";
+import { assert } from "chai";
 import { ethers } from "hardhat";
 import type {
   ReadWriteTier,
@@ -12,8 +12,6 @@ import {
   getEventArgs,
 } from "../../utils";
 import type { Contract } from "ethers";
-
-const { assert } = chai;
 
 const uninitializedReport =
   "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff";
