+++ test/Tier/TierReport.sol.ts
@@ -1,4 +1,4 @@
-import chai from "chai";
+import { assert } from "chai";
 import { ethers } from "hardhat";
 import type { ReadWriteTier } from "../../typechain/ReadWriteTier";
 import type { TierReportTest } from "../../typechain/TierReportTest";
@@ -6,8 +6,6 @@ import type { ReserveTokenTest } from "../../typechain/ReserveTokenTest";
 import { assertError, basicDeploy, zeroPad32, zeroPad4 } from "../../utils";
 import type { Contract } from "ethers";
 import type { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-
-const { assert } = chai;
 
 enum Tier {
   ZERO,
