+++ test/VM/CalculatorTestUnchecked.ts
@@ -1,5 +1,5 @@
 import * as Util from "../../utils";
-import chai from "chai";
+import { assert } from "chai";
 import { ethers } from "hardhat";
 import { concat } from "ethers/lib/utils";
 import { op } from "../../utils";
@@ -8,7 +8,6 @@ import type { Contract } from "ethers";
 import type { AllStandardOpsTest } from "../../typechain/AllStandardOpsTest";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-const { assert } = chai;
 
 const Opcode = Util.AllStandardOps;
