+++ test/VM/FixedPointMathsOpsTest.sol.ts
@@ -1,13 +1,11 @@
 import * as Util from "../../utils";
-import chai from "chai";
+import { assert } from "chai";
 import { ethers } from "hardhat";
 import { concat } from "ethers/lib/utils";
 import { op } from "../../utils";
 import type { Contract } from "ethers";
 
 import type { AllStandardOpsTest } from "../../typechain/AllStandardOpsTest";
-
-const { assert } = chai;
 
 const Opcode = Util.AllStandardOps;
