+++ utils/deploy/index.ts
@@ -1,5 +1,9 @@
 export * from "./basic";
+export * from "./claim";
 export * from "./combineTier";
+export * from "./emissions";
+export * from "./escrow";
 export * from "./redeemableERC20";
+export * from "./sale";
 export * from "./verify";
 export * from "./verifyTier";
