+++ utils/rainvm/index.ts
@@ -1,3 +1,4 @@
-export * from "./statebuilder";
 export * from "./ops";
+export * from "./sale";
+export * from "./statebuilder";
 export * from "./vm";
