+++ utils/rainvm/ops/allStandardOps.ts
