+++ utils/rainvm/sale.ts
@@ -1,0 +1,31 @@
+import { concat } from "ethers/lib/utils";
+import { AllStandardOps } from "./ops/allStandardOps";
+import { op } from "./vm";
+
+export const Opcode = AllStandardOps;
+
+export const afterBlockNumberSource = (constant: number): Uint8Array => {
+  // prettier-ignore
+  return concat([
+    // (BLOCK_NUMBER blockNumberSub1 gt)
+      op(Opcode.BLOCK_NUMBER),
+      op(Opcode.CONSTANT, constant),
+    op(Opcode.GREATER_THAN),
+  ]);
+};
+
+export const betweenBlockNumbersSource = (
+  vStart: Uint8Array,
+  vEnd: Uint8Array
+): Uint8Array => {
+  // prettier-ignore
+  return concat([
+        op(Opcode.BLOCK_NUMBER),
+        vStart,
+      op(Opcode.GREATER_THAN),
+        op(Opcode.BLOCK_NUMBER),
+        vEnd,
+      op(Opcode.LESS_THAN),
+    op(Opcode.EVERY, 2),
+  ])
+};
