+++ utils/tier/index.ts
@@ -1,9 +1,7 @@
 import { BigNumber } from "ethers";
 import { ethers } from "hardhat";
-import chai from "chai";
-import { max_uint256 } from "./constants";
-
-const { assert } = chai;
+import { assert } from "chai";
+import { max_uint256 } from "../constants";
 
 export const ALWAYS = 0;
 export const NEVER = max_uint256;
@@ -42,4 +40,19 @@ export function tierReport(report: string): number[] {
   }
 
   return parsedReport;
+}
+
+export function tierRange(startTier: number, endTier: number): number {
+  //   op_.val & 0x0f, //     00001111
+  //   op_.val & 0xf0, //     11110000
+
+  if (startTier < 0 || startTier > 8) {
+    throw new Error(`Invalid startTier ${startTier}`);
+  } else if (endTier < 0 || endTier > 8) {
+    throw new Error(`Invalid endTier ${endTier}`);
+  }
+  let range = endTier;
+  range <<= 4;
+  range += startTier;
+  return range;
 }
