+++ utils/types/sale.ts
@@ -1,0 +1,13 @@
+export enum Status {
+  PENDING,
+  ACTIVE,
+  SUCCESS,
+  FAIL,
+}
+
+export enum SaleStorage {
+  RemainingUnits,
+  TotalReserveIn,
+  Token,
+  Reserve,
+}
