+++ utils/types/tier.ts
@@ -1,0 +1,11 @@
+export enum Tier {
+  ZERO,
+  ONE,
+  TWO,
+  THREE,
+  FOUR,
+  FIVE,
+  SIX,
+  SEVEN,
+  EIGHT,
+}
